[#109] add visual value for field WIP-limits & move color picker to the folder shared
