+++ src/shared/colorPickerTooltip.js
